more review routes updated
